pyproject.toml
[tool.poetry]
name = "fantasy-football-scraper"
version = "0.1.0"
description = ""
authors = ["Connor Berg <bergconnor@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.38.1"
boto3 = "^1.24.56"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pylint = "^2.14.5"
bandit = "^1.7.4"
black = "^22.6.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.971"
safety = "^2.1.1"
radon = "^5.1.0"
xenon = "^0.9.0"
coverage = "^6.4.4"
pre-commit = "^2.20.0"

[tool.poetry.group.collect-statistics]
optional = true

[tool.poetry.group.collect-statistics.dependencies]
pandas = "^1.4.3"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
